pneumonia-main/pneumonia-main/requirements.txt
tensorflow
numpy
pillow
streamlit